+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/message/Catalog.kt
@@ -71,7 +71,10 @@ data class NotificationMessage(
         @SerializedName("minutes")                        val connectionTime:                Long,
         @SerializedName("channel_connections")            val numberOfConnections:           Int,
         @SerializedName("channel_upstream_connections")   val numberOfUpstreamConnections:   Int,
-        @SerializedName("channel_downstream_connections") val numberOfDownstreamConnections: Int
+        @SerializedName("channel_downstream_connections") val numberOfDownstreamConnections: Int,
+        @SerializedName("audio")                          val audio:                         Boolean?,
+        @SerializedName("video")                          val video:                         Boolean?,
+        @SerializedName("spotlight_id")                   val spotlightId:                   String?
 )
 
 data class PushMessage(