+++ CHANGES.md
@@ -33,6 +33,8 @@
 - SoraMediaOption に tcpCandidatePolicy を追加した
   - もともと内部的に用いていたオプションの格上げ
   - デフォルト値はこれまでと同様に ENABLED
+- `NotificationMessage` に `audio`, `video`, `spotlightId` を追加した
+  - すべてオプショナル
 
 ### CHANGE